Comments filter from most liked to least liked. Force one filter in URL.
