+++ app/main/templates/index.html
@@ -5,7 +5,7 @@
         {{ filter_form.hidden_tag() }}
         <p style="padding-left:5vw"> 
             {{ filter_form.filter.label }}
-            <select name="filter" id="filter">
+            <select name="filter" id="search-filter">
                 <option value="nwst" {% if filter_value == 'nwst' %} selected {% endif %}>Newest</option>
                 <option value="ldst" {% if filter_value == 'ldst' %} selected {% endif %}>Oldest</option>
                 <option value="mslk" {% if filter_value == 'mslk' %} selected {% endif %}>Most Liked</option>
@@ -26,12 +26,12 @@
             </p>
         {% endif %}
         <div id="toggle-comments-field-{{post.id}}" style="display: none;">
-            {% for comment in post.comments %}
+            {% for comment in sorted_comments %}
                 <p>
                     {% include '_comment.html' %}
                 </p>
             {% endfor %}
-        </div>
+        </div>              
     {% endfor %}
     {% if prev_url %}
         <a style="padding-left: 2vw;" href="{{ prev_url }}">Newer posts</a>
